Extending Mm, wip trying to save music to MIDI
Mm is rather insufficient when it comes to writing MIDI since it throws
away everything that is not `Note_on`, `Note_off` or control change,
which makes MIDI software like MuseScore reject my files. So my work so
far does not amount to much. Next steps are to actually keep all MIDI
events in order to write them to a file.